+++ signalfx/resource_signalfx_single_value_chart.go
@@ -54,6 +54,12 @@ func singleValueChartResource() *schema.Resource {
 				Description:  "How long (in seconds) to wait for late datapoints",
 				ValidateFunc: validation.IntBetween(0, 900),
 			},
+			"timezone": &schema.Schema{
+				Type:        schema.TypeString,
+				Optional:    true,
+				Default:     "UTC",
+				Description: "The property value is a string that denotes the geographic region associated with the time zone, (e.g. Australia/Sydney)",
+			},
 			"refresh_interval": &schema.Schema{
 				Type:        schema.TypeInt,
 				Optional:    true,
@@ -226,6 +232,12 @@ func getSingleValueChartOptions(d *schema.ResourceData) *chart.Options {
 		md := int32(val.(int) * 1000)
 		programOptions.MaxDelay = &md
 	}
+	if val, ok := d.GetOk("timezone"); ok {
+		if programOptions == nil {
+			programOptions = &chart.GeneralOptions{}
+		}
+		programOptions.Timezone = val.(string)
+	}
 	options.ProgramOptions = programOptions
 
 	if refreshInterval, ok := d.GetOk("refresh_interval"); ok {
@@ -319,6 +331,9 @@ func singlevaluechartAPIToTF(d *schema.ResourceData, c *chart.Chart) error {
 			if err := d.Set("max_delay", *options.ProgramOptions.MaxDelay/1000); err != nil {
 				return err
 			}
+		}
+		if err := d.Set("timezone", options.ProgramOptions.Timezone); err != nil {
+			return err
 		}
 	}